requirements.txt
asgiref==3.3.1
astroid==2.4.2
click==7.1.2
colorama==0.4.4
cycler==0.10.0
dj-database-url==0.5.0
Django==3.0.5
django-ckeditor==6.0.0
django-js-asset==1.2.2
isort==5.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
Pillow==8.1.0
protobuf==3.14.0
pylint==2.6.0
pylint-django==2.4.2
pylint-plugin-utils==0.6
pyparsing==2.4.7
pyperclip==1.8.1
python-dateutil==2.8.1
python-decouple==3.4
pytz==2020.5
screeninfo==0.6.7
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
Werkzeug==1.0.1
whitenoise==5.2.0
wrapt==1.12.1
